pyproject.toml
[tool.poetry]
name = "pyzantine"
version = "0.1.0"
description = ""
authors = ["Rich <richard.a.cairns@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.6b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
